src/codechef/medium/section14/addsquare/solution.go
package main

import (
	"bufio"
	"fmt"
	"os"
)

func main() {
	reader := bufio.NewReader(os.Stdin)
	firstLine := readNNums(reader, 4)
	W, H, N, M := firstLine[0], firstLine[1], firstLine[2], firstLine[3]
	X := readNNums(reader, N)
	Y := readNNums(reader, M)
	fmt.Println(solve(W, H, H, M, X, Y))
}

func readInt(bytes []byte, from int, val *int) int {
	i := from
	sign := 1
	if bytes[i] == '-' {
		sign = -1
		i++
	}
	tmp := 0
	for i < len(bytes) && bytes[i] >= '0' && bytes[i] <= '9' {
		tmp = tmp*10 + int(bytes[i]-'0')
		i++
	}
	*val = tmp * sign
	return i
}

func readNum(reader *bufio.Reader) (a int) {
	bs, _ := reader.ReadBytes('\n')
	readInt(bs, 0, &a)
	return
}

func readTwoNums(reader *bufio.Reader) (a int, b int) {
	res := readNNums(reader, 2)
	a, b = res[0], res[1]
	return
}

func readThreeNums(reader *bufio.Reader) (a int, b int, c int) {
	res := readNNums(reader, 3)
	a, b, c = res[0], res[1], res[2]
	return
}

func readNNums(reader *bufio.Reader, n int) []int {
	res := make([]int, n)
	x := 0
	bs, _ := reader.ReadBytes('\n')
	for i := 0; i < n; i++ {
		for x < len(bs) && (bs[x] < '0' || bs[x] > '9') && bs[x] != '-' {
			x++
		}
		x = readInt(bs, x, &res[i])
	}
	return res
}

func readUint64(bytes []byte, from int, val *uint64) int {
	i := from

	var tmp uint64
	for i < len(bytes) && bytes[i] >= '0' && bytes[i] <= '9' {
		tmp = tmp*10 + uint64(bytes[i]-'0')
		i++
	}
	*val = tmp

	return i
}

func solve(W, H, N, M int, X []int, Y []int) int {
	vert := NewBitSet(W + 1)

	for i := 0; i < N; i++ {
		vert.Set(X[i])
	}

	hori := NewBitSet(H + 1)
	rev := NewBitSet(H + 1)
	for i := 0; i < M; i++ {
		hori.Set(Y[i])
		rev.Set(H - Y[i])
	}

	vertDiff := NewBitSet(W + 1)

	for i := 0; i <= W; i++ {
		if vert.IsSet(i) {
			vertDiff = vertDiff.Or(vert.RightShift(i))
		}
	}

	horiDiff := NewBitSet(H + 1)

	for i := 0; i <= H; i++ {
		if hori.IsSet(i) {
			horiDiff = horiDiff.Or(hori.RightShift(i))
		}
	}
	var ans int

	for c := 0; c <= H; c++ {
		tmp := hori.RightShift(c)
		tmp = tmp.Or(rev.RightShift(H - c))
		tmp = tmp.Or(horiDiff)
		tmp = tmp.And(vertDiff)

		cnt := tmp.Count()

		if cnt > ans {
			ans = cnt
		}
	}

	return ans - 1
}

const BIT_SIZE = 64

type BitSet struct {
	arr []uint64
	sz  int
	n   int
}

func NewBitSet(n int) BitSet {
	sz := (n + BIT_SIZE - 1) / BIT_SIZE
	arr := make([]uint64, sz)
	return BitSet{arr, sz, n}
}

func (bs *BitSet) Set(i int) {
	arr := bs.arr
	if i >= bs.n {
		return
	}
	arr[i/BIT_SIZE] |= 1 << uint64(i%BIT_SIZE)
}

func (bs *BitSet) IsSet(i int) bool {
	arr := bs.arr
	if i >= bs.n {
		return false
	}
	return arr[i/BIT_SIZE]&(1<<uint64(i%BIT_SIZE)) > 0
}

func (bs *BitSet) Clear(i int) {
	if !bs.IsSet(i) {
		return
	}
	arr := bs.arr
	arr[i/BIT_SIZE] ^= 1 << uint64(i%BIT_SIZE)
}

func (bs BitSet) RightShift(x int) BitSet {
	arr := make([]uint64, len(bs.arr))

	a, b := x/BIT_SIZE, x%BIT_SIZE
	B := uint64(1<<uint64(b) - 1)
	var i int
	for a < bs.sz {
		cur := bs.arr[a]
		if b > 0 {
			cur >>= uint64(b)
			if a+1 < bs.sz {
				cur |= (bs.arr[a+1] & B) << (BIT_SIZE - uint64(b))
			}
		}
		arr[i] = cur
		i++
		a++
	}
	return BitSet{arr, bs.sz, bs.n}
}

func (this BitSet) And(that BitSet) BitSet {
	and := func(x, y uint64) uint64 {
		return x & y
	}
	return merge(this, that, and)
}

func (this BitSet) Or(that BitSet) BitSet {
	or := func(x, y uint64) uint64 {
		return x | y
	}
	return merge(this, that, or)
}

func merge(this, that BitSet, fn func(uint64, uint64) uint64) BitSet {
	arr := make([]uint64, this.sz)

	for i := 0; i < this.sz; i++ {
		arr[i] = fn(this.arr[i], that.arr[i])
	}

	return BitSet{arr, this.sz, this.n}
}

func (this BitSet) Count() int {
	var ans int

	for i := 0; i < this.sz; i++ {
		cur := this.arr[i]
		for cur > 0 {
			ans++
			cur -= cur & -cur
		}
	}
	return ans
}
